refactor: move more components to arrow-pg
