+++ Cargo.toml
@@ -16,6 +16,7 @@ documentation = "https://docs.rs/crate/datafusion-postgres/"
 [workspace.dependencies]
 arrow = "55"
 datafusion = { version = "47", default-features = false }
+futures = "0.3"
 pgwire = "0.30.2"
 postgres-types = "0.2"
 tokio = { version = "1", default-features = false }
