+++ arrow-pg/src/error.rs
@@ -1,0 +1,1 @@
+pub type ToSqlError = Box<dyn std::error::Error + Sync + Send>;
